requirements.txt
# Flask核心 (Python 3.8兼容版本)
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# 环境变量管理
python-dotenv==0.21.1

# HTTP请求
requests==2.28.2
urllib3==1.26.18
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# 日期时间处理
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0

# 开发依赖（可选）
# pytest==7.2.2
# pytest-flask==1.2.0

# 生产环境服务器（可选）
# gunicorn==21.2.0
# gevent==23.9.1